rename: common 모듈 패키지 위치 수정
