+++ core/common/src/main/java/com/tenday/core/common/enums/BadgeCode.kt
@@ -1,4 +1,4 @@
-package com.tenday.core.common.extensions
+package com.tenday.core.common.enums
 
 enum class BadgeCode {
     A, B, C, D, E, F;
