+++ core/common/src/main/java/com/tenday/core/common/extentions/Format.kt
@@ -1,4 +1,4 @@
-package com.tenday.core.common.enums
+package com.tenday.core.common.extentions
 
 fun Int.toData(): String {
     return "%,d".format(this)
