+++ core/designsystem/src/main/java/com/tenday/designsystem/components/profileCard/ProfileCard.kt
@@ -14,9 +14,9 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.tenday.core.common.extensions.BadgeCode
-import com.tenday.core.common.extensions.JobFamily
-import com.tenday.core.common.extensions.ProfileCode
+import com.tenday.core.common.enums.BadgeCode
+import com.tenday.core.common.enums.JobFamily
+import com.tenday.core.common.enums.ProfileCode
 import com.tenday.designsystem.R
 import com.tenday.designsystem.dimens.Dimens
 import com.tenday.designsystem.extentions.dashedFill
