+++ core/designsystem/src/main/java/com/tenday/designsystem/components/profileCard/ProfileCardFrames.kt
@@ -6,7 +6,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.painterResource
-import com.tenday.core.common.extensions.JobFamily
+import com.tenday.core.common.enums.JobFamily
 import com.tenday.designsystem.R
 
 @Composable
