+++ core/designsystem/src/main/java/com/tenday/designsystem/components/profileCard/ProfileCardName.kt
@@ -12,7 +12,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.tenday.core.common.extensions.BadgeCode
+import com.tenday.core.common.enums.BadgeCode
 import com.tenday.designsystem.R
 import com.tenday.designsystem.dimens.Dimens
 import com.tenday.designsystem.theme.HandsUpTypography
